refactor(seller): Resolver/Service DTO 전환 + 구 interface 정리
- Seller resolver 9종: @Args 인자 타입을 신규 DTO class 로 교체
  (ValidationPipe 가 검증). 구 interface 다중 import 라인을 개별 import 로 분해.
- Seller service 7종: import 경로를 dto/inputs/* 로 변경 (type import 유지).
- types/seller-input.type.ts (268 줄) 삭제 — 모든 정의가 DTO class 로 이전됨.
- 일부 spec 파일의 `as never` / `as unknown as X` 캐스트가 isolatedModules:false
  하에서 불필요해져 lint-staged 가 자동 정리.
- spec 1건의 startsAt/endsAt 입력을 string 에서 Date 로 정정
  (DTO 가 Date 만 허용; GraphQL DateTime scalar 가 변환 후 도달).

FE 영향: 정상 입력은 응답 동일. 잘못된 형식은 ValidationPipe 가
일관된 BAD_REQUEST 로 응답.
