+++ src/features/seller/resolvers/seller-order-mutation.resolver.ts
@@ -1,8 +1,8 @@
 import { UseGuards } from '@nestjs/common';
 import { Args, Mutation, Resolver } from '@nestjs/graphql';
 
+import { SellerUpdateOrderStatusInput } from '@/features/seller/dto/inputs/seller-update-order-status.input';
 import { SellerOrderService } from '@/features/seller/services/seller-order.service';
-import type { SellerUpdateOrderStatusInput } from '@/features/seller/types/seller-input.type';
 import type { SellerOrderSummaryOutput } from '@/features/seller/types/seller-output.type';
 import {
   CurrentUser,
